+ DFA : fix overflow in get_linear_bound when computing transition offset
state_id * layer_shape was computed as uint32_t * int (32-bit) before
assignment to size_t, causing overflow for large DFAs where state_id
exceeds 2^31 / layer_shape.

+++ src/DFA.cpp
@@ -570,7 +570,7 @@ const DFALinearBound& DFA::get_linear_bound() const
 		  bool local_accept_all = false;
 		  uint32_t local_bounds = 0;
 
-		  size_t offset = state_id * layer_shape;
+		  size_t offset = size_t(state_id) * size_t(layer_shape);
 		  for(size_t i = 0; i < layer_shape; ++i)
 		    {
 		      if(curr_transitions[offset + i] == 1)
